+++ changelog.md
@@ -1,5 +1,6 @@
 (07 mars 2017)
 - gestion des composants dans le styleguide 
+-  amélioration du fichier gulp
 
 # changelog v2.4.0 (23 nov 2016)